requirements.txt
# pip install -r requirements.txt
hl7apy
hexdump
flask 
flasgger
mysql-connector-python
requests
php
